Add pref listening to regex matcher (#123)
The multi-thread synchronisation is a bit messy, I wonder if there's
a better way of doing this? This is needed because webview might
keep loading things in the background - and we might then get mismatches
if the regex matcher has a different state than the trie matcher while
the regex matcher is reloading (whereas the trie matcher can
instantaneously switch).

It would probably be simpler to just pauseTimers()/resumeTimers()
on the webview.